Style(MessageQueue): Each class gets its own file.
Now each file in the webserver module gets its own file:
- Created messagequeue.h messagequeue.cpp.
- Move the MessageQueue class to messagequeue.h.
- Use of std:: prefix in standard C++ types instead of
"using namespace std;".

Fix(EXTRACFLAGS): Remove -DDISABLE_WEBSERVER from parameters.
